+++ test.php
@@ -177,18 +177,18 @@
 					$textY = $y1;
 
 					if ($i == ($cnt -1)) $textX = $x1 - ($textW);
-					if ($textY <= $paddingTop) $textY = $y1 + (__('@graph_font_size@') * 2);
+					if ($textY <= $paddingTop) $textY = $y1 + ((int)__('@graph_font_size@') * 2);
 					if ($i == 0) $textX = $paddingLeft;
 
 					if ($i == ($cnt -1) && $len !== $max) {
-						$lx1 = $paddingLeft;
-						$ly1 = $y1;
-						$lx2 = __('@graph_w@') - $paddingRight;
-						$ly2 = $y1;
-						imagedashedline($img, $lx1, $ly1, $lx2, $ly2, __('@graph_frame_color@'));
+						$lx1 = (int)$paddingLeft;
+						$ly1 = (int)$y1;
+						$lx2 = (int)__('@graph_w@') - $paddingRight;
+						$ly2 = (int)$y1;
+						imagedashedline($img, $lx1, $ly1, $lx2, $ly2, (int)__('@graph_frame_color@'));
 					}
 
-					imagettftext($img, __('@graph_font_size@'), 0, $textX, $textY, __('@graph_text_color@'), __('@graph_font@'), $text);
+					imagettftext($img, (int)__('@graph_font_size@'), 0, $textX, $textY, (int)__('@graph_text_color@'), __('@graph_font@'), $text);
 				}
 			}
 		}
